\name{SLAPE.all_genes_exonic_content_block_lengths_ensemble}
\alias{SLAPE.all_genes_exonic_content_block_lengths_ensemble}
\docType{data}
\title{
Genome-wide total exonic block lenghts
}
\description{
This object contains the total exnonic block lengths for all the genes.
}
\format{
 A genome-wide named numerical vector containing the total exonic block lengths for all the genes. Names of this vector are offical HUGO gene symbols.
 Please note that the name of this vector is \code{GECOBlengths}.
}
\details{
This vector can be updated by using the \code{SLAPE.compute_gene_exon_content_block_lengths} function.
}
\source{
}
\references{
}
\examples{
}
\keyword{datasets}
