\name{SLAPE.pathvis}
\alias{SLAPE.pathvis}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generatig a heatmap of the alteration matrix of a SLAPenriched pathway and barplots with statistical scores
}
\description{
    This function generates a pdf file containing a heatmap of the alteration matrix of a SLAPenriched pathway across the samples of the     analysed dataset, after permuting rows and columns to highlight trend of mutual exclusivity in the alteration-patterns.
    
    Additionally, it generates, a pdf file with three barplots indicating, respectively: (i) the number of mutated genes across samples;
    (ii) the probabilities of the pathway under consideration to be altered across samples, together with the expected number
    of samples with alteration in the pathway under consideration; (iii) The observed pathway alteration status across samples, together     with the observed total number of samples with alteration in the pathway under consideration.
    
}
\usage{
SLAPE.pathvis(EM, PFP, Id, prefName = NULL, PATH = "./", PATH_COLLECTION)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{EM}{
    A sparse binary matrix, or a sparse matrix with integer non-null entries. In this matrix the column names are sample identifiers, and     the row names official HUGO gene symbols. A non-zero entry in position \eqn{i,j} of this matrix indicates the presence of a somatic      mutations harbored by the \eqn{j-}sample in the \eqn{i-}gene. This matrix must be the same that has been inputted to the
       \code{\link{SLAPE.analyse}} function to produce the results in the \code{PFP} list.
}
  \item{PFP}{
    A list containing the SLAPenrich analysis results outputted by the \code{\link{SLAPE.analyse}} function while analysing
        the genomic dataset summarised by the \code{EM} genomic event matrix.
}
  \item{Id}{
    The index of the pathway for which the pdf files should be produced in the vectors/lists of the pathway collection set specified in 
    \code{PATH_COLLECTION}.
}
\item{prefName}{
   A string specifing the prefix to be added to the pdf file name (NULL by default).
}
  \item{PATH}{
        String specifiying the path of the pdf file to be created (including its name).
}
  \item{PATH_COLLECTION}{
    The pathway collection that has been tested against the \code{EM} dataset with the \code{\link{SLAPE.analyse}} function to produce           the \code{PFP} list of results.
}
}
\details{

}
\value{
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Francesco Iorio - iorio@ebi.ac.uk
}
\note{
This function makes use of the \code{\link{SLAPE.heuristic_mut_ex_sorting}} function to realise the mutual exclusivity sorting
of the pathway alteration matrix, and it is iteratively used by \code{\link{SLAPE.serialPathVis}} to generate pdf files with
heatmaps and barplots for all the SLAPenriched pathways found in a list of results generated by \code{\link{SLAPE.analyse}}.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
    \code{\link{SLAPE.analyse}}, \code{\link{SLAPE.serialPathVis}}
}
\examples{
    #Loading the Genomic Event data object derived from variants annotations
    #identified in 188 Lung Adenocarcinoma patients (Ding et al, 2008)
    data(LUAD_CaseStudy_updatedGS)

    #Loading KEGG pathway gene-set collection data object
    data(SLAPE.MSigDB_KEGG_hugoUpdated)

    #Loading genome-wide total exonic block lengths
    data(SLAPE.all_genes_exonic_content_block_lengths_ensemble)

    #Running SLAPenrich analysis
    PFPw<-SLAPE.analyse(EM = LUAD_CaseStudy_ugs,
                    PATH_COLLECTION = KEGG_PATH,
                    BACKGROUNDpopulation = rownames(LUAD_CaseStudy_ugs),
                    GeneLenghts = GECOBLenghts)

    #Generating a pdf file containing a heatmap of the mutual-exclusivity
    #sorted pathway alteration matrix, for an SLAPenriched pathway with
    #an exclusive coverage > 80\%, and barplots with statistical scores.
    #The pdf is saved in a file with \code{Example_} as prefix in its name,
    #in the current working directory.
    SLAPE.pathvis(EM = LUAD_CaseStudy_ugs,PFP = PFPw,
        Id = PFPw$pathway_id[which(PFPw$pathway_exclusiveCoverage>80)[1]],
        prefName = 'Example_',
        PATH = './',PATH_COLLECTION = KEGG_PATH)
    
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{Result-exploration}
